fixed getById

+++ backend/database.py
@@ -10,7 +10,6 @@ collection = database.Sessions
 async def fetch_one_session(muid):
     doc = await collection.find_one({"muid": muid})
     doc['_id'] = str(doc['_id'])
-    doc['id'] = str(doc['_id'])
     session_data = GetSessionModel(**doc)
     return session_data
 
@@ -19,8 +18,8 @@ async def fetch_all_sessions():
     sessions = []
     cursor = collection.find({})
     async for document in cursor:
+        print(document)
         document['_id'] = str(document['_id'])
-        document['id'] = str(document['_id'])
         session_data = GetSessionModel(**document)
         sessions.append(session_data)
     return sessions
@@ -32,7 +31,7 @@ async def create_session(session: Session):
     return document
 
 
-async def update_session(muid: int, session: Session):
+async def update_session(muid: str, session: Session):
     await collection.update_one({"muid": muid}, {"$set": {
         "device_type": session.device_type,
         "transfer_usd": session.transfer_usd,
@@ -44,6 +43,7 @@ async def update_session(muid: int, session: Session):
 
 async def delete_session(muid):
     doc = await collection.find_one({"muid": muid})
-    print(doc)
-    await collection.delete_one({"muid": muid})
-    return True
+    if doc:
+        await collection.delete_one({"muid": muid})
+        return True
+    return False
